+++ Test/UnitTests/TestScaffolder.cs
@@ -12,6 +12,7 @@ using Xunit;
 using Xunit.Abstractions;
 using Xunit.Extensions.AssertExtensions;
 
+#pragma warning disable EF1001 // Internal EF Core API usage.
 namespace Test.UnitTests
 {
     public class TestScaffolder 
@@ -46,6 +47,7 @@ namespace Test.UnitTests
 
             //VERIFY
             model.ShouldNotBeNull();
+            model.DefaultSchema.ShouldEqual("dbo");
         }
 
     }